+++ lib/bloom_remit_client/responses/create_payment_response.rb
@@ -3,10 +3,6 @@ module BloomRemitClient
 
     attribute :payment, Object, lazy: true, default: :default_payment
 
-    def success?
-      raw_response.success?
-    end
-
     private
 
     def default_payment